Rewrite value implementation
The original idea was to keep values in a single concrete type to enable
easy implementation early on, but this turns out to create quite messy code with
a lot of type switching (which harms performance), and required
allocation for all of the members (which also harms performance).

The real kicker, though, is that because the types are rather large, it
looks like they disturbed a nasty codepath in the Go runtime (upstream issue 24416)
which causes copy time to be awful. As we pass values around by value
everywhere, that hurts a lot: in sum10k, we ended up spending ~25% of
our runtime just copying around values.

Problem summarized, onto the solution: Switch value to an interface, and add a
concrete implementation of the interface per JS type.

Note that object is still structured similarly right now, and may also benefit
from the same treatment.

Benchmarks almost universally benefit, except from the String type (and
thus, the String object). I'm uncertain why it appears to be harmed by
this change, but it can likely be improved in the future (e.g. we aren't
interning any strings at all).

    benchmark                                     old ns/op     new ns/op     delta
    BenchmarkSimpleTypes/undefined-8              2.21          0.27          -87.78%
    BenchmarkSimpleTypes/null-8                   2.17          0.28          -87.10%
    BenchmarkSimpleTypes/bool-8                   7.13          0.29          -95.93%
    BenchmarkSimpleTypes/number-8                 15.0          0.28          -98.13%
    BenchmarkSimpleTypes/string_empty-8           11.1          0.28          -97.48%
    BenchmarkSimpleTypes/string_1c-8              11.9          0.27          -97.73%
    BenchmarkSimpleTypes/string_5c-8              14.2          0.28          -98.03%
    BenchmarkSimpleTypesFromVM/undefined-8        12692         10193         -19.69%
    BenchmarkSimpleTypesFromVM/null-8             10794         10473         -2.97%
    BenchmarkSimpleTypesFromVM/bool-8             13791         12196         -11.57%
    BenchmarkSimpleTypesFromVM/number-8           10592         10564         -0.26%
    BenchmarkSimpleTypesFromVM/string_empty-8     11114         13591         +22.29%
    BenchmarkSimpleTypesFromVM/string_1c-8        10427         12517         +20.04%
    BenchmarkSimpleTypesFromVM/string_5c-8        10935         13251         +21.18%
    BenchmarkObjectTypes/object-8                 12.6          2.75          -78.17%
    BenchmarkObjectTypes/bool-8                   20.6          3.54          -82.82%
    BenchmarkObjectTypes/number-8                 13.7          5.46          -60.15%
    BenchmarkObjectTypes/string_empty-8           71.7          5.71          -92.04%
    BenchmarkObjectTypes/string_1c-8              75.9          30.3          -60.08%
    BenchmarkObjectTypes/string_5c-8              65.8          34.6          -47.42%
    BenchmarkObjectTypes/function_call-8          12.6          3.25          -74.21%
    BenchmarkObjectTypes/function_ctor-8          12.6          3.30          -73.81%
    BenchmarkObjectTypes/function_both-8          12.8          3.25          -74.61%
    BenchmarkObjectFromSimpleType/string_5c-8     15094         16140         +6.93%
    BenchmarkVM/fib_it-8                          47070         37763         -19.77%
    BenchmarkVM/fib_rec-8                         149983        106784        -28.80%
    BenchmarkVM/sum10k-8                          2547995       1537682       -39.65%
    BenchmarkVM/sum10kf-8                         2594093       1668724       -35.67%