Switch property keys from string to value
And use it to specialize get() for strings to fetch characters

This is not done using getOwnProperty (and is thus somewhat wrong). We should
make get() a generic thing, and only specialize getOwnProperty I think.

Another potential improvement would be to avoid object coersion for strings
(specialize LOAD_INDEXED to work on a valueString directly).